+++ apps/web/src/graphql/queries/MeQuery.js
@@ -130,6 +130,7 @@ export default gql`
           showSuggestedSkills
           showWelcomePage
           layout
+          autoAddMembers
         }
         slug
         type