+++ apps/web/src/store/actions/fetchGroupSpaces.js
@@ -36,6 +36,12 @@ export default function fetchGroupSpaces (groupId) {
               accessibility
               requiredRoles
               paywall
+              settings {
+                allowGroupInvites
+                showWelcomePage
+                layout
+                autoAddMembers
+              }
               groupRoles {
                 items {
                   id