+++ apps/web/src/store/actions/fetchGroupsMenuData.js
@@ -77,6 +77,7 @@ query FetchGroupsMenuData (
         allowGroupInvites
         showWelcomePage
         layout
+        autoAddMembers
       }
       groupViews {
         items {
@@ -116,6 +117,12 @@ query FetchGroupsMenuData (
             accessibility
             requiredRoles
             paywall
+            settings {
+              allowGroupInvites
+              showWelcomePage
+              layout
+              autoAddMembers
+            }
             groupRoles {
               items {
                 id
@@ -200,6 +207,12 @@ query FetchGroupsMenuData (
           accessibility
           requiredRoles
           paywall
+          settings {
+            allowGroupInvites
+            showWelcomePage
+            layout
+            autoAddMembers
+          }
           groupRoles {
             items {
               id