+++ packages/graphql/fragments/membershipFieldsFragment.js
@@ -40,6 +40,7 @@ export default gql`
         showSuggestedSkills
         showWelcomePage
         layout
+        autoAddMembers
       }
       slug
       type